move classes into correct directory
