+++ poc/backend/app/src/main/java/ch/zhaw/studyflow/webserver/HttpResponse.java
@@ -5,7 +5,7 @@ import ch.zhaw.studyflow.webserver.contents.BodyContent;
 public interface HttpResponse extends HttpRequest {
     HttpResponse setResponseContent(BodyContent content);
 
-    HttpResponse setStatusCode(int i);
+    HttpResponse setStatusCode(HttpStatusCode statusCode);
 
     BodyContent getResponseContent();
 }
