+++ poc/backend/app/src/main/java/ch/zhaw/studyflow/webserver/WebServer.java
@@ -1,4 +1,5 @@
 package ch.zhaw.studyflow.webserver;
 
 public interface WebServer {
+    void start();
 }
