--- poc/backend/app/src/main/java/ch/zhaw/studyflow/webserver/java/HttpServerWebServer.java
@@ -1,6 +1,0 @@
-package ch.zhaw.studyflow.webserver.java;
-
-import ch.zhaw.studyflow.webserver.WebServer;
-
-public class HttpServerWebServer implements WebServer {
-}
