+++ poc/backend/app/src/main/java/ch/zhaw/studyflow/webserver/sun/SunHttpRequest.java
@@ -1,8 +1,8 @@
 package ch.zhaw.studyflow.webserver.sun;
 
-import ch.zhaw.studyflow.webserver.CookieContainer;
 import ch.zhaw.studyflow.webserver.HttpRequest;
 import ch.zhaw.studyflow.webserver.HttpResponse;
+import ch.zhaw.studyflow.webserver.cookies.CookieContainer;
 import com.sun.net.httpserver.HttpExchange;
 
 import java.nio.charset.Charset;
