+++ poc/backend/app/src/main/java/ch/zhaw/studyflow/webserver/sun/SunHttpResponse.java
