+++ poc/backend/app/src/main/java/ch/zhaw/studyflow/webserver/sun/SunHttpServerWebServer.java
@@ -1,0 +1,6 @@
+package ch.zhaw.studyflow.webserver.sun;
+
+import ch.zhaw.studyflow.webserver.WebServer;
+
+public class SunHttpServerWebServer implements WebServer {
+}
