--- poc/backend/src/main/java/ch/zhaw/studyflow/webserver/HttpResponse.java
@@ -1,11 +1,0 @@
-package ch.zhaw.studyflow.webserver;
-
-import ch.zhaw.studyflow.webserver.contents.BodyContent;
-
-public interface HttpResponse extends HttpRequest {
-    HttpResponse setResponseContent(BodyContent content);
-
-    HttpResponse setStatusCode(HttpStatusCode statusCode);
-
-    BodyContent getResponseContent();
-}
