--- poc/backend/src/main/java/ch/zhaw/studyflow/webserver/WebServer.java
@@ -1,5 +1,0 @@
-package ch.zhaw.studyflow.webserver;
-
-public interface WebServer {
-    void start();
-}
